balatro_rs::boss_blinds

Trait BossBlindInfo

pub trait BossBlindInfo {
    // Required methods
    fn name(&self) -> &'static str;
    fn description(&self) -> &'static str;
    fn base_score_requirement(&self) -> usize;
    fn reward_multiplier(&self) -> f64;
}

Trait for boss blind metadata and information Follows Interface Segregation Principle by focusing only on metadata concerns

Required Methods

fn name(&self) -> &'static str

Get the name of this boss blind

fn description(&self) -> &'static str

Get the description of what this boss blind does

fn base_score_requirement(&self) -> usize

Get the base score requirement for this boss blind

fn reward_multiplier(&self) -> f64

Get the reward multiplier for defeating this boss blind