balatro_rs::card_filter

Function filter_cards

pub fn filter_cards<'a, I, F>(
    cards: I,
    filter: &F,
    context: &FilterContext,
) -> Vec<&'a Card>
where
    I: Iterator<Item = &'a Card>,
    F: CardFilter,

Filter a collection of cards using the provided filter

Arguments

cards - Iterator over cards to filter

filter - The filter to apply

context - Context information for filtering

Returns

A vector containing only the cards that match the filter