balatro_rs::consumables

Enum ConsumableId

pub enum ConsumableId {

    TheFool,
    TheMagician,
    TheHighPriestess,
    TheEmperor,
    TheHierophant,
    TheEmpress,
    TheLovers,
    TheChariot,
    Strength,
    TheHermit,
    WheelOfFortune,
    Justice,
    TheHangedMan,
    Death,
    Temperance,
    TheDevil,
    TheTower,
    TheStar,
    TheMoon,
    TheSun,
    Judgement,
    TheWorld,
    Mercury,
    Venus,
    Earth,
    Mars,
    Jupiter,
    Saturn,
    Uranus,
    Neptune,
    Pluto,
    PlanetX,
    Ceres,
    Eris,
    Familiar,
    Grim,
    Incantation,
    Immolate,
    Ankh,
    DejaVu,
    Hex,
    Trance,
    Medium,
    Cryptid,
    TheSoul,
    BlackHole,
    TarotPlaceholder,
    PlanetPlaceholder,
    SpectralPlaceholder,

}

Identifier for all consumable cards in the game This will be extended as consumable implementations are added

Variants

TheFool

The Fool - Creates last Joker used this round if possible

TheMagician

The Magician - Enhances 2 selected cards to Lucky Cards

TheHighPriestess

The High Priestess - Creates up to 2 Planet Cards

TheEmperor

The Emperor - Creates up to 2 Tarot Cards

TheHierophant

The Hierophant - Enhances 2 selected cards to Bonus Cards

TheEmpress

The Empress - Enhances 2 selected cards to Mult Cards

TheLovers

The Lovers - Enhances 1 selected card to Wild Card

TheChariot

The Chariot - Enhances 1 selected card to Steel Card

Strength

Strength - Increases rank of up to 2 selected cards by 1

TheHermit

The Hermit - Gain $20 money

WheelOfFortune

Wheel of Fortune - 1 in 4 chance to add Foil, Holographic, or Polychrome edition

Justice

Justice (XI) - Enhances 1 selected card to Glass Card

TheHangedMan

The Hanged Man (XII) - Destroys up to 2 selected cards

Death

Death (XIII) - Select 2 cards, convert left card to right card

Temperance

Temperance (XIV) - Gives total sell value of all Jokers as money

TheDevil

The Devil (XV) - Enhances 1 selected card to Gold Card

TheTower

The Tower (XVI) - Enhances 1 selected card to Stone Card

TheStar

The Star (XVII) - Converts up to 3 selected cards to Diamonds

TheMoon

The Moon (XVIII) - Converts up to 3 selected cards to Clubs

TheSun

The Sun (XIX) - Converts up to 3 selected cards to Hearts

Judgement

Judgement (XX) - Creates a random Joker card

TheWorld

The World (XXI) - Converts up to 3 selected cards to Spades

Mercury

Mercury - Levels up Pair

Venus

Venus - Levels up Two Pair

Earth

Earth - Levels up Full House

Mars

Mars - Levels up Three of a Kind

Jupiter

Jupiter - Levels up Straight

Saturn

Saturn - Levels up Straight

Uranus

Uranus - Levels up Two Pair

Neptune

Neptune - Levels up Straight Flush

Pluto

Pluto - Levels up High Card

PlanetX

Planet X - Levels up Five of a Kind

Ceres

Ceres - Levels up Flush House

Eris

Eris - Levels up Flush Five

Familiar

Familiar - Destroys 1 random card, add 3 random Enhanced face cards to deck

Grim

Grim - Destroys 1 random card, add 2 random Enhanced Aces to deck

Incantation

Incantation - Destroys 1 random card, add 4 random Enhanced numbered cards to deck

Immolate

Immolate - Destroys 5 random cards in hand, gain $20

Ankh

Ankh - Create copy of random Joker, destroy all other Jokers

DejaVu

Deja Vu - Add Red Seal to 1 selected card

Hex

Hex - Add Polychrome to random Joker, destroy all other Jokers

Trance

Trance - Add Blue Seal to 1 selected card

Medium

Medium - Add Purple Seal to 1 selected card

Cryptid

Cryptid - Create 2 copies of 1 selected card

TheSoul

The Soul - Creates a Legendary Joker (must be room)

BlackHole

Black Hole - Upgrade every hand type by 1 level

TarotPlaceholder

Placeholder for future Tarot card implementations

PlanetPlaceholder

Placeholder for future Planet card implementations

SpectralPlaceholder

Placeholder for future Spectral card implementations

Implementations

impl ConsumableId

pub fn all() -> Vec<ConsumableId>

Get all available consumable IDs

pub fn consumable_type(&self) -> ConsumableType

Get the consumable type for this ID

pub fn tarot_cards() -> Vec<ConsumableId>

Get all Tarot cards

pub fn planet_cards() -> Vec<ConsumableId>

Get all Planet cards

pub fn spectral_cards() -> Vec<ConsumableId>

Get all Spectral cards