balatro_rs

Module joker

Re-exports

pub use conditional::ConditionalJoker;

pub use conditional::JokerCondition;

pub use traits::JokerGameplay;

pub use traits::JokerIdentity;

pub use traits::JokerLifecycle;

pub use traits::JokerModifiers;

pub use traits::JokerState as JokerStateTrait;

pub use traits::ProcessContext;

pub use traits::ProcessResult;

pub use traits::Rarity;

pub use compat::Joker as OldJoker;

pub use compat::Jokers;

Modules

basic_additive_mult_jokers

Basic Additive Mult Jokers implementation

basic_economy_jokers

Basic Economy Jokers implementation

compat

conditional

four_fingers

FourFingers Joker implementation

hand_composition_jokers

multiplicative_jokers

Multiplicative jokers implementation

resource_chips_jokers

Resource-Based Chips Jokers implementation

scaling_xmult_jokers

Scaling XMult Jokers implementation

traits

New trait definitions for the Joker system

Structs

GameContext

Context provided to joker methods for accessing game state

JokerEffect

Effect that a joker can apply to modify game state.

Enums

Categories

JokerId

Enum representing all 150 joker identifiers

JokerRarity

Joker rarity levels

Traits

Joker

Core trait that all jokers must implement.