balatro_rs::joker

Module basic_additive_mult_jokers

Basic Additive Mult Jokers implementation

This module implements jokers that provide additive mult bonuses (+X Mult). These jokers add fixed amounts of mult to the score under various conditions.

Structs

BasicJoker

Basic Joker - +4 Mult per hand

EvenStevenJoker

Even Steven - Even cards (2, 4, 6, 8, 10) give +4 Mult when scored

HalfJoker

Half Joker - +20 Mult if played hand has 4 or fewer cards

ScholarJoker

Scholar - Aces give +20 Chips and +4 Mult when scored

WalkieJoker

Walkie Talkie - Each played 10 or 4 gives +10 Chips and +4 Mult when scored

Functions

create_basic_joker

Factory functions for creating basic additive mult jokers

create_even_steven_joker

create_half_joker

create_scholar_joker

create_walkie_joker