balatro_rs::joker

Module resource_chips_jokers

Resource-Based Chips Jokers implementation

This module implements jokers that provide chip bonuses based on game resources like money, discards, cards in deck, etc.

Structs

BannerJoker

Banner joker - +30 chips per remaining discard

BlueJoker

Blue joker - +2 chips per remaining card in deck

BullJoker

Bull joker - +2 chips per $1 owned

ScaryFaceJoker

Scary Face joker - +30 chips when face cards are scored

StoneJoker

Stone joker - +25 chips per Stone card in deck

Functions

create_banner_joker

Factory functions for creating resource-based chips jokers

create_blue_joker

create_bull_joker

create_scary_face_joker

create_stone_joker