balatro_rs::joker

Module retrigger_jokers

Retrigger Jokers implementation

This module implements jokers that retrigger card effects. These jokers cause cards to be scored multiple times, multiplying their effects.

Structs

DuskJoker

Dusk - Retriggers all played cards in the last hand of the round

HangingChadJoker

Hanging Chad - Retriggers the first card scored twice

SeltzerJoker

Seltzer - Retriggers all played cards for 10 hands, then self-destructs

SockAndBuskinJoker

Sock and Buskin - Retriggers all face cards

Functions

create_dusk_joker

Factory functions for creating retrigger jokers

create_hanging_chad_joker

create_seltzer_joker

create_sock_and_buskin_joker