balatro_rs::joker::traits

Trait JokerLifecycle

pub trait JokerLifecycle: Send + Sync {
    // Provided methods
    fn on_purchase(&mut self) { ... }
    fn on_sell(&mut self) { ... }
    fn on_destroy(&mut self) { ... }
    fn on_round_start(&mut self) { ... }
    fn on_round_end(&mut self) { ... }
    fn on_joker_added(&mut self, _other_joker_type: &str) { ... }
    fn on_joker_removed(&mut self, _other_joker_type: &str) { ... }
}

Trait for joker lifecycle management.

This trait handles the lifecycle events of a joker, from purchase through gameplay to sale or destruction.

Provided Methods

fn on_purchase(&mut self)

Called when the joker is purchased from the shop.

fn on_sell(&mut self)

Called when the joker is sold.

fn on_destroy(&mut self)

Called when the joker is destroyed.

fn on_round_start(&mut self)

Called at the start of each round.

fn on_round_end(&mut self)

Called at the end of each round.

fn on_joker_added(&mut self, _other_joker_type: &str)

Called when another joker is added to the collection.

fn on_joker_removed(&mut self, _other_joker_type: &str)

Called when another joker is removed from the collection.