balatro_rs

Module shop

Re-exports

pub use packs::PackType;

Modules

generation

packs

Structs

EnhancedShop

Enhanced shop with support for multiple item types and weighted generation

ItemWeights

Weights for different item types in shop generation

JokerGenerator

Pack

A booster pack containing multiple items

Shop

ShopSlot

Individual slot in the enhanced shop

Enums

ConsumableType

Types of consumable cards available in the shop

ShopItem

Represents all possible items that can appear in the shop.

SlotModifier

Special modifiers that can affect shop slots

VoucherId

Voucher identifiers for shop vouchers

Traits

PackSelector

Interface for pack selection and opening mechanics

RerollMechanics

Interface for reroll mechanics

ShopGenerator

Enhanced shop trait for generating shop contents with weighted randomization and support for various item types including jokers, consumables, vouchers, and packs.