balatro_rs::shop

Module packs

Structs

DefaultPackGenerator

Default pack generator implementation

OpenPackState

State for an opened pack that player is currently choosing from

Pack

A booster pack containing multiple options for the player to choose from

PackOption

An individual option within a pack

Enums

PackType

Types of booster packs available in the shop

Traits

PackGenerator

Trait for pack generation and management