balatro_rs::shop::packs

Trait PackGenerator

pub trait PackGenerator {
    // Required methods
    fn generate_pack(
        &self,
        pack_type: PackType,
        game: &Game,
    ) -> Result<Pack, GameError>;
    fn is_pack_available(&self, pack_type: PackType, game: &Game) -> bool;
    fn available_pack_types(&self, game: &Game) -> Vec<PackType>;
}

Trait for pack generation and management

Required Methods

fn generate_pack( &self, pack_type: PackType, game: &Game, ) -> Result<Pack, GameError>

Generate a pack of the specified type

fn is_pack_available(&self, pack_type: PackType, game: &Game) -> bool

Check if a pack type is available for purchase

fn available_pack_types(&self, game: &Game) -> Vec<PackType>

Get all available pack types for current game state