balatro_rs::shop

Trait PackSelector

pub trait PackSelector {
    // Required methods
    fn select_from_pack(&self, pack: &Pack, game: &Game) -> Vec<ShopItem>;
    fn max_selections(&self, pack_type: PackType) -> usize;
    fn is_valid_selection(&self, pack: &Pack, selected: &[ShopItem]) -> bool;
}

Interface for pack selection and opening mechanics

Required Methods

fn select_from_pack(&self, pack: &Pack, game: &Game) -> Vec<ShopItem>

Select items from a pack based on pack type and game state

fn max_selections(&self, pack_type: PackType) -> usize

Get the maximum number of items that can be selected from a pack

fn is_valid_selection(&self, pack: &Pack, selected: &[ShopItem]) -> bool

Check if a pack selection is valid