balatro_rs::skip_tags

Module utility_tags

Utility Skip Tags

Implementation of the 4 utility category skip tags:

Double: Duplicates the next selected tag

Boss: Re-rolls the next Boss Blind

Orbital: Upgrades a random poker hand by 3 levels

Juggle: Adds +3 hand size for next round only

Structs

BossTag

Boss Tag - Re-rolls the next Boss Blind If Director’s Cut has been used, this will also consume its one re-roll option

DoubleTag

Double Tag - Gives a copy of the next Tag selected (excluding Double Tags) Can be stacked, with each addition creating one additional copy of a tag

JuggleTag

Juggle Tag - Adds +3 Hand Size for the next round only Can be stacked multiple times, with each tag adding an additional +3

OrbitalTag

Orbital Tag - Upgrades a specified random Poker Hand by three levels This can include a secret hand if it has been played in the current run