balatro_rs::consumables

Trait Consumable

pub trait Consumable:
    Send
    + Sync
    + Debug {
    // Required methods
    fn consumable_type(&self) -> ConsumableType;
    fn can_use(&self, game_state: &Game, target: &Target) -> bool;
    fn use_effect(
        &self,
        game_state: &mut Game,
        target: Target,
    ) -> Result<(), ConsumableError>;
    fn get_description(&self) -> String;
    fn get_target_type(&self) -> TargetType;
    fn get_effect_category(&self) -> ConsumableEffect;

    // Provided methods
    fn name(&self) -> &'static str { ... }
    fn description(&self) -> &'static str { ... }
    fn cost(&self) -> usize { ... }
    fn apply_effect(&self, game: &mut Game) -> bool { ... }
}

Core trait that all consumable types must implement Enhanced version with target validation and effect categorization

Required Methods

fn consumable_type(&self) -> ConsumableType

Get the consumable type category

fn can_use(&self, game_state: &Game, target: &Target) -> bool

Check if this consumable can be used with the given target in the current game state

fn use_effect( &self, game_state: &mut Game, target: Target, ) -> Result<(), ConsumableError>

Apply the effect of this consumable to the game state Future versions will support async for animations

fn get_description(&self) -> String

Get the description of what this consumable does

fn get_target_type(&self) -> TargetType

Get the type of target this consumable requires

fn get_effect_category(&self) -> ConsumableEffect

Get the effect category for this consumable

Provided Methods

fn name(&self) -> &'static str

Get the name of this consumable

fn description(&self) -> &'static str

Get the description as static str (legacy)

fn cost(&self) -> usize

Get the cost of this consumable in the shop

fn apply_effect(&self, game: &mut Game) -> bool

Legacy apply effect method for backward compatibility