balatro_rs

Module joker

Re-exports

pub use conditional::ConditionalJoker;

pub use conditional::JokerCondition;

pub use traits::JokerGameplay;

pub use traits::JokerIdentity;

pub use traits::JokerLifecycle;

pub use traits::JokerModifiers;

pub use traits::JokerState as JokerStateTrait;

pub use traits::ProcessContext;

pub use traits::ProcessResult;

pub use traits::Rarity;

pub use compat::Joker as OldJoker;

pub use compat::Jokers;

pub use advanced_conditions::AdvancedCondition;

pub use advanced_conditions::AdvancedConditionBuilder;

pub use advanced_conditions::AdvancedConditionEvaluator;

pub use advanced_conditions::AdvancedEvaluationContext;

pub use advanced_conditions::ConditionCache;

pub use advanced_conditions::GameHistory;

pub use advanced_traits::AdvancedJokerFramework;

pub use advanced_traits::AdvancedJokerGameplay;

pub use advanced_traits::AdvancedJokerIdentity;

pub use advanced_traits::EnhancedJoker;

pub use advanced_traits::EnhancedJokerBuilder;

pub use advanced_traits::EvaluationCost;

pub use advanced_traits::GameEvent;

pub use advanced_traits::InternalJokerState;

pub use advanced_traits::JokerProcessor;

pub use compatibility_bridge::CompatibilityBridge;

pub use compatibility_bridge::LegacyJokerAdapter;

pub use compatibility_bridge::MixedJokerCollection;

pub use scaling_chips_jokers::ArrowheadJoker;

pub use scaling_chips_jokers::CastleJoker;

pub use scaling_chips_jokers::HikerJoker;

pub use scaling_chips_jokers::OddToddJoker;

pub use scaling_chips_jokers::ScholarJoker;

pub use scaling_chips_jokers::StuntmanJoker;

pub use scaling_chips_jokers::WeeJoker;

Modules

advanced_conditions

Advanced Joker Condition Framework

advanced_traits

Advanced Joker Traits with Enhanced Condition Support

basic_economy_jokers

Basic Economy Jokers implementation

basic_xmult_jokers

Basic XMult Jokers implementation

compat

compatibility_bridge

Backward Compatibility Bridge for Advanced Joker Framework

conditional

four_fingers

FourFingers Joker implementation

hand_composition_jokers

multiplicative_jokers

Multiplicative jokers implementation

resource_chips_jokers

Resource-Based Chips Jokers implementation

retrigger_jokers

Retrigger Jokers implementation

scaling_additive_mult_jokers

Scaling Additive Mult Jokers implementation

scaling_chips_jokers

Scaling Chips Jokers for Issue #644

scaling_xmult_jokers

Scaling XMult Jokers implementation

traits

New trait definitions for the Joker system

Structs

GameContext

Context provided to joker methods for accessing game state

JokerEffect

Effect that a joker can apply to modify game state.

Enums

Categories

JokerId

Enum representing all 150 joker identifiers

JokerRarity

Joker rarity levels

Traits

Joker

Core trait that all jokers must implement.