balatro_rs::skip_tags

Trait SkipTag

pub trait SkipTag:
    Debug
    + Send
    + Sync {
    // Required methods
    fn id(&self) -> SkipTagId;
    fn name(&self) -> &'static str;
    fn description(&self) -> &'static str;
    fn effect_type(&self) -> TagEffectType;
    fn rarity(&self) -> TagRarity;
    fn stackable(&self) -> bool;
    fn activate(&self, context: SkipTagContext) -> SkipTagResult;

    // Provided methods
    fn selectable(&self) -> bool { ... }
    fn can_activate(&self, _context: &SkipTagContext) -> bool { ... }
    fn can_apply(&self, _game: &Game) -> bool { ... }
    fn apply_effect(&self, _game: &mut Game) -> TagEffectResult { ... }
}

Core trait for all skip tags

Required Methods

fn id(&self) -> SkipTagId

Get the tag ID

fn name(&self) -> &'static str

Get the display name

fn description(&self) -> &'static str

Get the description

fn effect_type(&self) -> TagEffectType

Get the effect type

fn rarity(&self) -> TagRarity

Get the rarity (affects skip chance)

fn stackable(&self) -> bool

Can this tag be stacked?

fn activate(&self, context: SkipTagContext) -> SkipTagResult

Activate the skip tag effect

Provided Methods

fn selectable(&self) -> bool

Can this tag be selected (some are automatic)

fn can_activate(&self, _context: &SkipTagContext) -> bool

Check if this tag can be activated in the given context

fn can_apply(&self, _game: &Game) -> bool

Check if this tag can be applied to the given game state

fn apply_effect(&self, _game: &mut Game) -> TagEffectResult

Apply the tag effect to the game state